Code Review Changes - Small optimizations and tweaks.
