+++ src/Events/Category_Colors/Migration/Worker.php
@@ -46,16 +46,16 @@ class Worker {
 	];
 
 	/**
-	 * Executor constructor.
-	 * This ensures the migration data is validated before execution,
-	 * preventing accidental execution with invalid data.
+	 * Allows for enabling Dry run mode.
 	 *
-	 * @since TBD
+	 * @param bool $dry_run Whether the worker should be in Dry Run mode.
 	 *
-	 * @param bool $dry_run Whether to run in dry-run mode.
+	 * @return self
 	 */
-	public function __construct( bool $dry_run = false ) {
+	public function set_dry_run( bool $dry_run = false ): self {
 		$this->dry_run = $dry_run;
+
+		return $this;
 	}
 
 	/**
@@ -71,18 +71,17 @@ class Worker {
 	 * @return void
 	 */
 	public function execute(): void {
-		$start_time = $this->start_timer();
+		$start_time = microtime( true );
 		if ( Status::$validation_completed !== $this->get_migration_status()['status'] ) {
 			$this->log_message( 'info', 'Validation not completed. Running validation before execution.', [], 'Worker' );
 
-			$validator = new Validator();
-
-			if ( ! $validator->validate() ) {
+			if ( ! tribe( Validator::class )->validate() ) {
 				$this->log_message( 'error', 'Validation failed. Migration execution stopped.', [], 'Worker' );
 				$this->update_migration_status( Status::$execution_failed ); // Mark execution as failed.
 
 				do_action( 'tec_events_category_colors_migration_runner_end', false );
 				$this->log_elapsed_time( 'Execution', $start_time );
+
 				return;
 			}
